JwtChanges

+++ loan/src/main/java/loan/loan/service/LoanServiceImpl.java
@@ -8,10 +8,17 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import loan.loan.datamodel.UserD;
 import loan.loan.exception.LoanNotFound;
 import loan.loan.exception.LoanServiceException;
+import loan.loan.exception.UserNotFound;
+import loan.loan.exception.UserUnAuthorized;
 import loan.loan.model.Loan;
+import loan.loan.model.MyUserDetails;
+import loan.loan.model.User;
 import loan.loan.repo.LoanRepo;
+import loan.loan.restclient.LoginClient;
+import loan.loan.security.JwtUtil;
 
 
 @Service
@@ -22,19 +29,59 @@ public class LoanServiceImpl implements LoanService {
 	@Autowired
 	LoanRepo loanRepo;
 	
+	@Autowired
+	LoginClient loginClient;
+	
+
+	@Autowired
+	JwtUtil jwtUtil;
+	
 	@Override
 	public Loan addLoan(Loan loan) {
-		try {
-			return loanRepo.save(loan);
-		} catch (Exception e) {
+
+		Loan loanObj = null;
+		
+		try{
+		UserD user = loginClient.getLogin(loan.getUser().getUserid());
+		if (user == null) {
+			throw new UserNotFound("User Not Found");
+		}
+		User userObj = new User(user.getUid(), user.getUsername(), user.getUserpwd(), user.getUsertype(),user.getJwtToken());
+		MyUserDetails userDetails = new MyUserDetails(userObj);
+		if (user.getUsertype().equals("ADMIN") && jwtUtil.validateToken(user.getJwtToken(), userDetails)) {
+			loanObj= loanRepo.save(loan);
+		}
+		 else {
+			throw new UserUnAuthorized("User Un Authorized");
+		}
+		}
+		catch(UserUnAuthorized e){
+			logger.error("Exceprtion occured: User Un Authorized" );
+			throw e;
+		}
+		catch(UserNotFound e){
+			logger.error("Exceprtion occured: User Not Found" );
+			throw e;
+		}
+		catch (Exception e) {
 			logger.error("Exceprtion occured: LoanServiceException" );
 			throw new LoanServiceException("Something went wrong..!");
 		}
+		return loanObj;
 	}
 	@Override
 	public Loan getLoan(int id)  {
-		Loan loanObj=null;
+		Loan loanObj = null;
 		try {
+		UserD user = loginClient.getLogin(id);
+		if (user == null) {
+			throw new UserNotFound("User Not Found");
+		}
+		User userObj = new User(user.getUid(), user.getUsername(), user.getUserpwd(), user.getUsertype(),
+				user.getJwtToken());
+		MyUserDetails userDetails = new MyUserDetails(userObj);
+		if (jwtUtil.validateToken(user.getJwtToken(), userDetails)) {
+		
 			Optional<Loan> optional = loanRepo.findById(id)	;
 			if (optional.isPresent()) {
 				loanObj=optional.get();
@@ -43,6 +90,17 @@ public class LoanServiceImpl implements LoanService {
 				
 				throw new LoanNotFound("Loan Not found");
 			}
+		}
+		else {
+			throw new UserUnAuthorized("User Un Authorized");
+		}
+		}
+		catch(UserUnAuthorized e){
+			logger.error("Exceprtion occured: User Un Authorized" );
+			throw e;
+		}
+		catch(UserNotFound e){
+			logger.error("Exceprtion occured: User Not found" );
 		}
 		catch(LoanNotFound e){
 			logger.error("Exceprtion occured: Loan Not found" );
@@ -57,9 +115,18 @@ public class LoanServiceImpl implements LoanService {
 
 	@Override
 	public Loan updateLoan(Loan loan) {
-		Optional<Loan> optional = loanRepo.findById(loan.getLoanno())	;
 		Loan loanObj = null;
 		try {
+		UserD user = loginClient.getLogin(loan.getUser().getUserid());
+		if (user == null) {
+			throw new UserNotFound("User Not Found");
+
+		}
+		User userObj = new User(user.getUid(), user.getUsername(), user.getUserpwd(), user.getUsertype(),
+				user.getJwtToken());
+		MyUserDetails userDetails = new MyUserDetails(userObj);
+		if (user.getUsertype().equals("ADMIN") && jwtUtil.validateToken(user.getJwtToken(), userDetails)) {
+		Optional<Loan> optional = loanRepo.findById(loan.getLoanno())	;
 			if (optional.isPresent()) {
 				loanObj = optional.get();
 				loanObj.setLoanno(loan.getLoanno());
@@ -76,6 +143,17 @@ public class LoanServiceImpl implements LoanService {
 				throw new LoanNotFound("Loan Not found");
 			}
 		}
+		else {
+			throw new UserUnAuthorized("User Un Authorized");
+		}
+		}
+		catch(UserUnAuthorized e){
+			logger.error("Exceprtion occured: User Un Authorized" );
+			throw e;
+		}
+		catch(UserNotFound e){
+			logger.error("Exceprtion occured: User Not found" );
+		}
 		catch(LoanNotFound e){
 				logger.error("Exceprtion occured: Loan Not found" );
 				throw e;
@@ -90,13 +168,39 @@ public class LoanServiceImpl implements LoanService {
 	@Override
 	public boolean deleteLoan(int id) {
 		boolean result =false;
+		
 		try {
+		UserD user = loginClient.getLogin(id);
+		if (user == null) {
+			throw new UserNotFound("User Not Found");
+
+		}
+		User userObj = new User(user.getUid(), user.getUsername(), user.getUserpwd(), user.getUsertype(),
+				user.getJwtToken());
+		MyUserDetails userDetails = new MyUserDetails(userObj);
+		if (user.getUsertype().equals("ADMIN") && jwtUtil.validateToken(user.getJwtToken(), userDetails)) {
+		
+	
 			loanRepo.deleteById(id);
 			return true;
-		} catch (Exception e) {
+		}else {
+			throw new UserUnAuthorized("User Un Authorized");
+			
+		}
+		}
+		catch(UserUnAuthorized e){
+			logger.error("Exceprtion occured: User Un Authorized" );
+			throw e;
+		}
+		catch(UserNotFound e){
+			logger.error("Exceprtion occured: User Not found" );
+		}
+		
+		catch (Exception e) {
 			logger.error("Exceprtion occured: Loan Not found" );
 			throw new LoanNotFound("Loan Not found");
 		}
+		return result;
 		
 	}
 	@Override
